pyproject.toml
[project]
name = "arb-bot"
version = "0.1.0"
description = "Arbitrage finder between Polymarket and Kalshi prediction markets"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "rapidfuzz>=3.5.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "tenacity>=8.2.0",
    "cryptography>=41.0.0",
]

[project.scripts]
arb-bot = "src.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
